test_runner: move global setup watch mode test to single file
